third_party/blink/web_tests/virtual/customizable-select-disabled/external/wpt/html/semantics/forms/the-select-element/customizable-select/selectedcontent.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] The <selectedcontent> element should reflect the HTML contents of the selected <option>.
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'innerHTML')"
[FAIL] When there are multiple <selectedcontent> elements, only the one in tree order should be kept up to date.
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'textContent')"
[FAIL] <seletedcontent> behavior in disconnected <select>.
  assert_equals: <selectedcontent> should be updated when <select> is connected to the document. expected "two" but got ""
Harness: the test ran to completion.
